pyproject.toml
[project]
name = "shensu"
version = "0.1.0"
description = "申诉提交脚本"
authors = [{ name = "yml" }]
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "cryptography>=42.0",
    "python-dotenv>=1.0",
]

[project.scripts]
submit-plea = "main:main"

[tool.uv]
index-url = "https://pypi.org/simple"

[tool.uv.sources]
wechat_tool = { path = "src/wechat_tool" }
